requirements.txt
snowflake-connector-python==3.15.0
snowflake-cli==3.9.1
altair~=5.5.0
snowflake-snowpark-python
snowflake-connector-python[pandas]
streamlit
python-dotenv==1.0.1